Moving data structures to mocks
